:zap: chore: abstract components
