+++ web/src/App.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import 'leaflet/dist/leaflet.css';
 import './styles/global.css';
 
 import Routes from './routes';
