+++ web/src/pages/CreateOrphanage.tsx
@@ -1,36 +1,17 @@
 import React from 'react';
 import { Map, Marker, TileLayer } from 'react-leaflet';
-import L from 'leaflet';
-import { useHistory } from 'react-router-dom';
+import { FiPlus } from 'react-icons/fi';
 
-import { FiArrowLeft, FiPlus } from 'react-icons/fi';
+import Sidebar from '../components/Sidebar';
 
-import mapMarkerImg from '../images/map-marker.svg';
+import mapIcon from '../utils/mapIcon';
 
 import '../styles/pages/create-orphanage.css';
 
-const happyMapIcon = L.icon({
-  iconUrl: mapMarkerImg,
-
-  iconSize: [58, 68],
-  iconAnchor: [29, 68],
-  popupAnchor: [0, -60],
-});
-
 export default function CreateOrphanage() {
-  const { goBack } = useHistory();
-
   return (
     <div id="page-create-orphanage">
-      <aside>
-        <img src={mapMarkerImg} alt="Happy" />
-
-        <footer>
-          <button type="button" onClick={goBack}>
-            <FiArrowLeft size={24} color="#FFF" />
-          </button>
-        </footer>
-      </aside>
+      <Sidebar />
 
       <main>
         <form className="create-orphanage-form">
@@ -48,7 +29,7 @@ export default function CreateOrphanage() {
 
               <Marker
                 interactive={false}
-                icon={happyMapIcon}
+                icon={mapIcon}
                 position={[-27.2092052, -49.6401092]}
               />
             </Map>
