+++ web/src/pages/OrphanagesMap.tsx
@@ -1,23 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { FiArrowRight, FiPlus } from 'react-icons/fi';
 import { Map, TileLayer, Marker, Popup } from 'react-leaflet';
-import Leaflet from 'leaflet';
+import { FiArrowRight, FiPlus } from 'react-icons/fi';
 
-import 'leaflet/dist/leaflet.css';
+import mapIcon from '../utils/mapIcon';
 
 import mapMarkerImg from '../images/map-marker.svg';
 
 import '../styles/pages/orphanages-map.css';
-
-const mapIcon = Leaflet.icon({
-  iconUrl: mapMarkerImg,
-
-  iconSize: [58, 68],
-  iconAnchor: [29, 68],
-
-  popupAnchor: [170, 2],
-});
 
 function OrphanagesMap() {
   return (
