+++ web/src/styles/pages/orphanage.css
@@ -3,46 +3,6 @@
   min-height: 100vh;
 }
 
-#page-orphanage aside {
-  position: fixed;
-  height: 100%;
-  padding: 32px 24px;
-  background: linear-gradient(329.54deg, #15B6D6 0%, #15D6D6 100%);
-
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-}
-
-#page-orphanage aside img {
-  width: 48px;
-}
-
-#page-orphanage aside footer a,
-#page-orphanage aside footer button {
-  width: 48px;
-  height: 48px;
-
-  border: 0;
-
-  background: #12AFCB;
-  border-radius: 16px;
-
-  cursor: pointer;
-
-  transition: background-color 0.2s;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-#page-orphanage aside footer a:hover,
-#page-orphanage aside footer button:hover {
-  background: #17D6EB;
-}
-
 #page-orphanage main {
   flex: 1;
 }
@@ -51,8 +11,8 @@
   width: 700px;
   margin: 64px auto;
 
-  background: #FFFFFF;
-  border: 1px solid #D3E2E5;
+  background: #ffffff;
+  border: 1px solid #d3e2e5;
   border-radius: 20px;
 
   overflow: hidden;
@@ -66,12 +26,12 @@
 
 .orphanage-details .images {
   display: grid;
-  grid-template-columns: repeat(6 ,1fr);
+  grid-template-columns: repeat(6, 1fr);
   column-gap: 16px;
 
   margin: 16px 40px 0;
 }
- 
+
 .orphanage-details .images button {
   border: 0;
   height: 88px;
@@ -80,7 +40,7 @@
   border-radius: 20px;
   overflow: hidden;
   outline: none;
-  
+
   opacity: 0.6;
 }
 
@@ -92,7 +52,6 @@
   width: 100%;
   height: 88px;
   object-fit: cover;
-  
 }
 
 .orphanage-details .orphanage-details-content {
@@ -100,7 +59,7 @@
 }
 
 .orphanage-details .orphanage-details-content h1 {
-  color: #4D6F80;
+  color: #4d6f80;
   font-size: 54px;
   line-height: 54px;
   margin-bottom: 8px;
@@ -108,14 +67,14 @@
 
 .orphanage-details .orphanage-details-content p {
   line-height: 28px;
-  color: #5C8599;
+  color: #5c8599;
   margin-top: 24px;
 }
 
 .orphanage-details .orphanage-details-content .map-container {
   margin-top: 64px;
-  background: #E6F7FB;
-  border: 1px solid #B3DAE2;
+  background: #e6f7fb;
+  border: 1px solid #b3dae2;
   border-radius: 20px;
 }
 
@@ -126,12 +85,15 @@
 
 .orphanage-details .orphanage-details-content .map-container footer a {
   line-height: 24px;
-  color: #0089A5;
+  color: #0089a5;
   text-decoration: none;
 }
 
-.orphanage-details .orphanage-details-content .map-container .leaflet-container {
-  border-bottom: 1px solid #DDE3F0;
+.orphanage-details
+  .orphanage-details-content
+  .map-container
+  .leaflet-container {
+  border-bottom: 1px solid #dde3f0;
   border-radius: 20px;
 }
 
@@ -139,14 +101,14 @@
   width: 100%;
   height: 1px;
   border: 0;
-  background: #D3E2E6;
+  background: #d3e2e6;
   margin: 64px 0;
 }
 
 .orphanage-details .orphanage-details-content h2 {
   font-size: 36px;
   line-height: 46px;
-  color: #4D6F80;
+  color: #4d6f80;
 }
 
 .orphanage-details .orphanage-details-content .open-details {
@@ -169,15 +131,18 @@
 }
 
 .orphanage-details .orphanage-details-content .open-details div.hour {
-  background: linear-gradient(149.97deg, #E6F7FB 8.13%, #FFFFFF 92.67%);
-  border: 1px solid #B3DAE2;
-  color: #5C8599;
+  background: linear-gradient(149.97deg, #e6f7fb 8.13%, #ffffff 92.67%);
+  border: 1px solid #b3dae2;
+  color: #5c8599;
 }
 
-.orphanage-details .orphanage-details-content .open-details div.open-on-weekends {
-  background: linear-gradient(154.16deg, #EDFFF6 7.85%, #FFFFFF 91.03%);
-  border: 1px solid #A1E9C5;
-  color: #37C77F;
+.orphanage-details
+  .orphanage-details-content
+  .open-details
+  div.open-on-weekends {
+  background: linear-gradient(154.16deg, #edfff6 7.85%, #ffffff 91.03%);
+  border: 1px solid #a1e9c5;
+  color: #37c77f;
 }
 
 .orphanage-details .orphanage-details-content button.contact-button {
@@ -187,9 +152,9 @@
   height: 64px;
   border: 0;
   cursor: pointer;
-  background: #3CDC8C;
+  background: #3cdc8c;
   border-radius: 20px;
-  color: #FFFFFF;
+  color: #ffffff;
   font-weight: 800;
 
   display: flex;
@@ -204,5 +169,5 @@
 }
 
 .orphanage-details .orphanage-details-content button.contact-button:hover {
-  background: #36CF82;
+  background: #36cf82;
 }
